David asset switch (#5)
* Fixed Creating Asset bug

Created a new Global to account for a new struct and adjusted the script naming and files to follow the standard of previous files.

* Add tyLibMtl support

first iteration on adding support for tyLibMaterial. Also started the ground work (previous push?) to add multimaterial saving per object...

* Updating the Material Publishing

It's not 4th of July here in Brazeeell! :P

* Update PublishModulesStruct.ms

Some bug fixing, code reduction (some un-needed routines) and more scenarios implemented

* Some more updates on the materials of AssetSwitch

Ok... I guess we're VERY close to a finished state on it. I found a nasty bug with tyLibMtl that might or might not be a deal breaker (TBH). Let's talk about it!

* Update PublishCenterUI.ms

Minor bug fix on VRayMeshExport introduced when I eliminated the "nodeAssetForPublish" variable in favor of "nodeAsset"

* Goal: Remove HOLD

Overall cleanup, little IF THEN refactoring for more direct result from condition clauses and variables renaming.

* Getting rid of HOLD/FETCH

UNDO is a very hacky way to replace HOLD/FETCH. Investigate an alternative by using MESHER and/or TYMESHER to combine the meshes instead without collapsing them and later delete the MESHER/TYMESHER object after VRayExport

* Bug Fixes and VRayMesh Combined Meshes

Fixed a bug brought up by Lisa and also finished implementing a way to export VRMeshes independently

* Bug fixing on Material Load

Bad spring splicing was returning a wrong filename for the matmap file

* Update PublishModulesStruct.ms

Fixed a bug that would crash the script if the user tried to create a vrmesh before creating a max publish of an asset

* First Pass VRayProxy

First working build of Vray Proxy, both in combined meshes or individual meshes

* Version 1.7

Deep cleanup of unused functions and UI items (still some more to do...) and some bug fixing here and there. Probably safe for deploying (fingers crossed!)

* Bug fixing!

Bug fixing in the case a VrayMesh is loaded but its published parent and replacement nodes are deleted or changed names. It still publishes but writes a warning log to make the concession explicit

* Bug fixing

Fixed bugs on the order of structs creation and material loading

* Fixed a VRayPart bug

Now when trying to load a Vray Parts asset, if the substitution asset is not found the script will load the original transform matrix in parent space and convert to the new world space based on the parent's current position. Nice!!!  😃 🍰

* Refactored the VRayMeshParts import

Now the VrayMesh parts importer works exactly the same as the Max importer, because it is in the end importing a Max file with all the Vray mesh parts already lodaded into it. The  other method to load vray mesh parts (by storing their transforms and loading them on switch) is still present in case it comes in handy in the future (AS 2.0)

* Update PythonTemplatePanelsUI.py

Python updates on the code to make it more streamlined and with a better UI

* Python interface for AS 2.0

UI interface for the next iteration of Asset Switch

* Lil advances on AS 2.0

* Update AssetSwitchUI.py

* Update AssetSwitchUI.py

---------

Co-authored-by: Make Visual ( Aaron Dabelow ) <60147676+makevisual@users.noreply.github.com>