Reorg UI, Anim load bugfixes, Switch enabled fix
Reorganized the Asset Switch UI to better reflect the actual positioning of buttons, checkboxes, dropdownlists, etc.

Also fixed the publishing of Animation files, making sure objects with Point Cache modifiers are not included, and objects with Attachment Constraint are not included too.

Finally created a function to properly enable and disable the "Switch Asset" button, to make it enabled only when a proper object/material/animation group is checked AND a valid file is selected on the dropdownlist