RC  v 1.7.4, with some showstopping bugfixes for the animation publish workflow and also some improvements on the UI responsiveness to the current state of the tool
* Partial bug fix Animation Workflow

The Save/Load animation has a severe limitation in regard to nodes that have an Attachment Constraint controller. This renders the process of save incomplete (because the Attachment Constraint keys are not saved) and makes the Load process destroy the asset's pose or animation.

A fix would involve :
1 - either check each added node to the save load routine for an attachment constraint controller (definitive fix but limits the scenarios where the save load is useful)
2 - On node collecting prior to save and load animation routine, exclude any HELPER objects, since those are the ones where (most of the time) an Attachment Constraint controller is added to. This would not avoid errors in case an attachment controller is added to other objects that are not Helpers though.

* Relocated the NAME BAKE function

Realocated the name bake function to make sure everytime the user hits the PUBLISH button, AS bakes the asset name on the object's UserProperties. Before, it only baked if the user Published a Max Asset.

* Reorg UI, Anim load bugfixes, Switch enabled fix

Reorganized the Asset Switch UI to better reflect the actual positioning of buttons, checkboxes, dropdownlists, etc.

Also fixed the publishing of Animation files, making sure objects with Point Cache modifiers are not included, and objects with Attachment Constraint are not included too.

Finally created a function to properly enable and disable the "Switch Asset" button, to make it enabled only when a proper object/material/animation group is checked AND a valid file is selected on the dropdownlist

* Better commenting and cleanup

* small version # update

* Little bug fix for Anim Publishing

The tool was lacking a folder creation on publishing.

* Some bugfixing and function scope adjusting